Add files via upload
We are provided with historical sales data for 45 Walmart stores located in different regions. Each store contains a number of departments, and you are tasked with predicting the department-wide sales for each store. We will use various Machine Learning algorithms and try to find the best model that predicts the sales values as closely as possible.