Add files via upload
Write a JavaScript program to generate a random number and store it in a variable. The program then takes input from the user to tell them whether the guess was correct or not. 100 - The number of Guesses is the score of the user. The program is expected to terminate once the number is guessed